# swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "varsion": "0.0.1",
    "title": "DAS SDL API server",
    "description": "DAS SDL 서버 API swagger 문서",
    "version": "1.0.0"
  },
  "host": "localhost:3200",
  "basePath": "/das",
...
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "token",
      "description": "유저token"
    }
  },
  "paths": {
...
    "/comm/notiInfo": {
      "get": {
        "tags": [
          "comm"
        ],
        "summary": "공지사항 조회",
        "description": "사용자가 공지사항을 조회한다",
        "parameters": [
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "resultCode": {
                  "type": "number",
                  "example": 200
                },
                "resultMsg": {
                  "type": "string",
                  "example": "공지사항 조회 성공"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "noti_id" : {
                        "type" : "number",
                        "example" : 15
                      },
                      "notice_title" : {
                        "type" : "string",
                        "example" : "공지사항제목"
                      },
                      "rgst_user_id" : {
                        "type" : "number",
                        "example" : 6
                      },
                      "user_name" : {
                        "type" : "string",
                        "example" : "홍길동"
                      },
                      "rgst_dtm" : {
                        "type" : "string",
                        "example" : "2024-08-29 16:58:09"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": [
              {
                "token": ""
              }
            ]
          }
        ]
      }
    },
    "/comm/addNoti": {
      "post": {
        "tags": [
          "comm"
        ],
        "summary": "공지사항 등록",
        "description": "사용자가 공지사항을 등록한다",
        "parameters" : [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "notice_title": { 
                  "type": "string",
                  "example" : "공지사항_1", 
                  "required" : true
                },
                "notice_description" : {
                  "type": "string",
                  "example" : "공지사항 내용 어쩌구저쩌구", 
                  "required" : true
                },
                "rgst_user_id" : {
                  "type" : "number",
                  "example" : 1,
                  "required" : true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "resultCode": {
                  "type": "number",
                  "example": 200
                },
                "resultMsg": {
                  "type": "string",
                  "example": "공지사항 등록 성공"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": [
              {
                "token": ""
              }
            ]
          }
        ]
      }
    },
    "/comm/detail": {
      "get": {
        "tags": [
          "comm"
        ],
        "summary": "공지사항 상세 조회",
        "description": "사용자가 공지사항의 상세화면을 호출한다",
        "parameters": [
          {
            "name": "noti_id",
            "description": "공지사항 아이디",
            "in": "query",
            "type": "number",
            "example" : 1,
            "required" : true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "resultCode": {
                  "type": "number",
                  "example": 200
                },
                "resultMsg": {
                  "type": "string",
                  "example": "공지사항 상세 조회 성공"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "noti_id" : {
                      "type": "number",
                      "example" : 1
                    },
                    "notice_title": { 
                      "type": "string",
                      "example" : "공지사항_1"
                    },
                    "notice_description" : {
                      "type": "string",
                      "example" : "공지사항 내용 어쩌구저쩌구"
                    },
                    "rgst_user_id" : {
                      "type" : "number",
                      "example" : 1
                    },
                    "user_name" : {
                      "type" : "string",
                      "example" : "홍길동"
                    },
                    "rgst_dtm" : {
                      "type" : "string",
                      "example" : "2024-08-26 16:01:40"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": [
              {
                "token": ""
              }
            ]
          }
        ]
      }
    }
  }
}